fix(Repository): null pointer in `Repository` module
In this commit, fix two kind of null pointer dereference. First,
`RepoService` relies on `UserDAO` and `RepoDAO`. However, in `SetupRepositoryRoutes`
only set the `RepoDAO`. And `UserDAO` is an empty pointer, which leads to
null pointer dereference when try to use it to access database.

Second, in `UserDAO.Create`, dereference a pointer whose value is `nil`.