fix: git push problem using go-git
There is a serious and strange problem before this commit. Every time
you create a repo for a user, in `gitolite` it shows everything is ok.
However, when you try to clone the repo, it refuse you to clone,
actually, it didn't add the new-created repos! However if you just run
`git commit --amend` with nothing changed, and then `git push`, you will
be able to clone the new-created repo before. This is a strange problem
and now i have not found the exact reason. For the primary analysis,
the problem should be located at the `push` of `go-git`. Because, after
using `go-git` to add and commit changes, we can push it manually, and
it works right. I have try to use python and `GitPython` to push, and it
works. Besides, before launch `main.go`, if run `eval "$(ssh-agent -s)"`
and `ssh-add ~/.ssh/id_rsa` subsequently, it works ok too. So I suppose
it must be incorrect `push setting` in `go-git`. I can't explain for
more details, but whatever it has been solved even if in an inelegant
way.