fix(gitolite): bugs in uploading sshkey
Before this commit, there is a serious problem when user upload their
ssh-key. The file used to save user ssh-key will not have `\n` at the
end of ssh-key string. This caused gitolite unable to recognize ssh-key.
So when user try to clone Repos from gitolite, they can't be verified
and will be asked for a password. And user also has not the access to
repository that belongs to himself.

The reason is that, when write user ssh-key into user file, only write
ssh-key itself. So there is no `\n` in the end of the ssh-key string. It
confuses gitolite.